+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Mail:
 ownCloud (now Nextcloud):
 
 * ownCloud is replaced with Nextcloud 10.0.5.
+* Fixed an error in Owncloud/Nextcloud setup not updating domain when changing hostname.
 
 Control Panel/Management:
 
@@ -19,6 +20,7 @@ Control Panel/Management:
 * Fix an error in the control panel related to IPv6 addresses.
 * TLS certificates for internationalized domain names can now be provisioned from Let's Encrypt automatically.
 * Download management web assets (jQuery/Bootstrap) to the static web root directory.
+
 
 v0.22 (April 2, 2017)
 ---------------------